+++ packages/lambda-tiler/src/cli/render.tile.ts
@@ -1,13 +1,13 @@
-import { ConfigProviderMemory, getAllImagery } from '@basemaps/config';
-import { initConfigFromPath } from '@basemaps/config/build/json/tiff.config.js';
-import { Epsg, ImageFormat, Nztm2000QuadTms } from '@basemaps/geo';
+import { ConfigProviderMemory } from '@basemaps/config';
+import { initConfigFromPaths } from '@basemaps/config/build/json/tiff.config.js';
+import { ImageFormat, Nztm2000QuadTms } from '@basemaps/geo';
 import { LogConfig, setDefaultConfig } from '@basemaps/shared';
 import { fsa } from '@chunkd/fs';
 import { LambdaHttpRequest, LambdaUrlRequest, UrlEvent } from '@linzjs/lambda';
 import { Context } from 'aws-lambda';
 import { TileXyzRaster } from '../routes/tile.xyz.raster.js';
 
-const target = `/home/blacha/tmp/basemaps/`;
+const target = `/home/blacha/tmp/basemaps`;
 const tile = { z: 18, x: 126359, y: 137603 };
 const tileMatrix = Nztm2000QuadTms;
 
@@ -15,12 +15,11 @@ async function main(): Promise<void> {
   const log = LogConfig.get();
   const provider = new ConfigProviderMemory();
   setDefaultConfig(provider);
-  const tileSet = await initConfigFromPath(provider, target);
+  const { tileSet, imagery } = await initConfigFromPaths(provider, [target]);
 
   if (tileSet.layers.length === 0) throw new Error('No imagery found in path: ' + target);
   log.info({ tileSet: tileSet.name, layers: tileSet.layers.length }, 'TileSet:Loaded');
-  const imagery = await getAllImagery(provider, tileSet.layers, [Epsg.Nztm2000, Epsg.Google]);
-  for (const im of imagery.values()) {
+  for (const im of imagery) {
     log.info({ imagery: im.uri, title: im.title, tileMatrix: im.tileMatrix, files: im.files.length }, 'Imagery:Loaded');
   }
   const request = new LambdaUrlRequest({ headers: {} } as UrlEvent, {} as Context, log) as LambdaHttpRequest;
@@ -33,6 +32,7 @@ async function main(): Promise<void> {
   });
 
   await fsa.write(`./${tile.z}_${tile.x}_${tile.y}.png`, Buffer.from(res.body, 'base64'));
+  log.info({ path: `./${tile.z}_${tile.x}_${tile.y}.png` }, 'Tile:Write');
 }
 
 main();