+++ packages/landing/src/components/layout.header.tsx
@@ -19,6 +19,7 @@ export class Header extends Component<unknown, HeaderState> {
 
   componentDidMount(): void {
     this._events.push(Config.map.on('change', () => this.forceUpdate()));
+    this._events.push(Config.map.on('filter', () => this.renderLinksTiles()));
 
     // If individual layers are on, we need the layer info to determine if they can use NZTM2000Quad WMTS
     Config.map.layers.then((layers) => this.setState({ layers }));
@@ -178,12 +179,26 @@ Your Service/App URL:
 
     const children: ReactNode[] = [];
     if (projections.has(EpsgCode.Nztm2000)) {
-      const nztmTileUrl = Config.map.toTileUrl(MapOptionType.Wmts, Nztm2000QuadTms);
+      const nztmTileUrl = Config.map.toTileUrl(
+        MapOptionType.Wmts,
+        Nztm2000QuadTms,
+        undefined,
+        undefined,
+        undefined,
+        Config.map.filter.date,
+      );
       children.push(<Copyable key="NZTM2000Quad" header="WMTS: NZTM2000Quad" value={nztmTileUrl} />);
     }
 
     if (projections.has(EpsgCode.Google)) {
-      const googleTileUrl = Config.map.toTileUrl(MapOptionType.Wmts, GoogleTms);
+      const googleTileUrl = Config.map.toTileUrl(
+        MapOptionType.Wmts,
+        GoogleTms,
+        undefined,
+        undefined,
+        undefined,
+        Config.map.filter.date,
+      );
       const googleXyzTileUrl = Config.map.toTileUrl(MapOptionType.TileRaster, GoogleTms);
       children.push(<Copyable key="WebMercatorQuad" header="WMTS: WebMercatorQuad" value={googleTileUrl} />);
       children.push(<Copyable key="XYZ" header="XYZ" value={googleXyzTileUrl} />);