+++ packages/landing/src/components/map.tsx
@@ -1,12 +1,15 @@
 import { GoogleTms } from '@basemaps/geo';
-import maplibre from 'maplibre-gl';
+import maplibre, { RasterLayerSpecification } from 'maplibre-gl';
 import { Component, ReactNode } from 'react';
 import { MapAttribution } from '../attribution.js';
 import { Config } from '../config.js';
 import { getTileGrid, locationTransform } from '../tile.matrix.js';
-import { WindowUrl } from '../url.js';
+import { MapOptionType, WindowUrl } from '../url.js';
+import { DateRange } from './daterange.js';
 import { Debug } from './debug.js';
 import { MapSwitcher } from './map.switcher.js';
+
+const LayerFadeTime = 750;
 
 /**
  * Map loading in maplibre is weird, the on('load') event is different to 'loaded'
@@ -80,12 +83,64 @@ export class Basemaps extends Component<unknown, { isLayerSwitcherEnabled: boole
   updateStyle = (): void => {
     this.ensureGeoControl();
     const tileGrid = getTileGrid(Config.map.tileMatrix.identifier);
-    const style = tileGrid.getStyle(Config.map.layerId, Config.map.style);
+    const style = tileGrid.getStyle(Config.map.layerId, Config.map.style, undefined, Config.map.filter.date);
     this.map.setStyle(style);
 
     if (Config.map.tileMatrix !== GoogleTms) this.map.setMaxBounds([-180, -85.06, 180, 85.06]);
     else this.map.setMaxBounds();
+    // TODO check and only update when Config.map.layer changes.
     this.forceUpdate();
+  };
+
+  updateVisibleLayers = (newLayers: string): void => {
+    if (!Config.map.visibleLayers) Config.map.visibleLayers = newLayers;
+    if (newLayers !== Config.map.visibleLayers) {
+      Config.map.visibleLayers = newLayers;
+      const newStyleId =
+        `${Config.map.styleId}` +
+        `_after=${Config.map.filter.date.after?.slice(0, 4)}` +
+        `&before=${Config.map.filter.date.before?.slice(0, 4)}`;
+      if (this.map.getSource(newStyleId) == null) {
+        this.map.addSource(newStyleId, {
+          type: 'raster',
+          tiles: [
+            WindowUrl.toTileUrl({
+              urlType: MapOptionType.TileRaster,
+              tileMatrix: Config.map.tileMatrix,
+              layerId: Config.map.layerId,
+              config: Config.map.config,
+              date: Config.map.filter.date,
+            }),
+          ],
+          tileSize: 256,
+        });
+        this.map.addLayer({
+          id: newStyleId,
+          type: 'raster',
+          source: newStyleId,
+          paint: { 'raster-opacity': 0 },
+        });
+        this.map.moveLayer(newStyleId); // Move to front
+        this.map.setPaintProperty(newStyleId, 'raster-opacity-transition', { duration: LayerFadeTime });
+        this.map.setPaintProperty(newStyleId, 'raster-opacity', 1);
+      }
+    }
+  };
+
+  removeOldLayers = (): void => {
+    const filteredLayers = this.map
+      .getStyle()
+      .layers.filter((layer) => layer.id.startsWith(Config.map.styleId)) as RasterLayerSpecification[];
+    // The last item in the array is the top layer, we pop that to ensure it isn't removed
+    filteredLayers.pop();
+    for (const layer of filteredLayers) {
+      this.map.setPaintProperty(layer.id, 'raster-opacity-transition', { duration: LayerFadeTime });
+      this.map.setPaintProperty(layer.id, 'raster-opacity', 0);
+      setTimeout(() => {
+        this.map.removeLayer(layer.id);
+        this.map.removeSource(layer.source);
+      }, LayerFadeTime);
+    }
   };
 
   componentDidMount(): void {
@@ -116,12 +171,14 @@ export class Basemaps extends Component<unknown, { isLayerSwitcherEnabled: boole
     }
 
     this.map.on('render', this.onRender);
+    this.map.on('idle', this.removeOldLayers);
     onMapLoaded(this.map, () => {
       this._events.push(
         Config.map.on('location', this.updateLocation),
         Config.map.on('tileMatrix', this.updateStyle),
         Config.map.on('layer', this.updateStyle),
         Config.map.on('bounds', this.updateBounds),
+        Config.map.on('visibleLayers', this.updateVisibleLayers),
       );
 
       this.updateStyle();
@@ -144,6 +201,9 @@ export class Basemaps extends Component<unknown, { isLayerSwitcherEnabled: boole
       <div style={{ flex: 1, position: 'relative' }}>
         <div id="map" style={{ width: '100%', height: '100%' }} />
         {Config.map.isDebug ? <Debug map={this.map} /> : undefined}
+        {Config.map.isDebug && !Config.map.debug['debug.screenshot'] && Config.map.debug['debug.date'] ? (
+          <DateRange />
+        ) : undefined}
         {isLayerSwitcherEnabled ? <MapSwitcher /> : undefined}
       </div>
     );