+++ packages/landing/src/config.map.ts
@@ -1,7 +1,8 @@
 import { Epsg, EpsgCode, GoogleTms, Nztm2000QuadTms, Nztm2000Tms, TileMatrixSet, TileMatrixSets } from '@basemaps/geo';
 import { Emitter } from '@servie/events';
-import maplibregl, { LngLatBoundsLike } from 'maplibre-gl';
-import { DebugState, DebugDefaults, ConfigDebug } from './config.debug.js';
+import { LngLatBoundsLike } from 'maplibre-gl';
+import { DateRangeState } from './components/daterange.js';
+import { ConfigDebug, DebugDefaults, DebugState } from './config.debug.js';
 import { Config } from './config.js';
 import { locationTransform } from './tile.matrix.js';
 import { ensureBase58, MapLocation, MapOptionType, WindowUrl } from './url.js';
@@ -13,12 +14,18 @@ const DefaultCenter: Record<string, MapLocation> = {
   [Nztm2000QuadTms.identifier]: { lat: -41.88999621, lon: 174.04924373, zoom: 3 },
 };
 
+export interface Filter {
+  date: DateRangeState;
+}
+
 export interface MapConfigEvents {
   location: [MapLocation];
   tileMatrix: [TileMatrixSet];
   layer: [string, string | null | undefined];
   bounds: [LngLatBoundsLike];
+  filter: [Filter];
   change: null;
+  visibleLayers: [string];
 }
 
 export class MapConfig extends Emitter<MapConfigEvents> {
@@ -27,6 +34,8 @@ export class MapConfig extends Emitter<MapConfigEvents> {
   tileMatrix: TileMatrixSet = GoogleTms;
   config: string | null;
   debug: DebugState = { ...DebugDefaults };
+  visibleLayers: string;
+  filter: Filter = { date: { after: undefined, before: undefined } };
 
   private _layers: Promise<Map<string, LayerInfo>>;
   get layers(): Promise<Map<string, LayerInfo>> {
@@ -76,6 +85,11 @@ export class MapConfig extends Emitter<MapConfigEvents> {
     return `${this.layerKey}::${this.tileMatrix.identifier}`;
   }
 
+  /** Used as source and layer id in the Style JSON for a given layer ID */
+  get styleId(): string {
+    return `basemaps-${Config.map.layerId}`;
+  }
+
   updateFromUrl(search: string = window.location.search): void {
     const urlParams = new URLSearchParams(search);
     const style = urlParams.get('s') ?? urlParams.get('style');
@@ -121,8 +135,9 @@ export class MapConfig extends Emitter<MapConfigEvents> {
     layerId = this.layerId,
     style = this.style,
     config = this.config,
+    date = this.filter.date,
   ): string {
-    return WindowUrl.toTileUrl(urlType, tileMatrix, layerId, style, config);
+    return WindowUrl.toTileUrl({ urlType, tileMatrix, layerId, style, config, date });
   }
 
   getLocation(map: maplibregl.Map): MapLocation {
@@ -148,6 +163,14 @@ export class MapConfig extends Emitter<MapConfigEvents> {
   setTileMatrix(tms: TileMatrixSet): void {
     if (this.tileMatrix.identifier === tms.identifier) return;
     this.emit('tileMatrix', this.tileMatrix);
+    this.emit('change');
+  }
+
+  setFilterDateRange(after: string | undefined, before: string | undefined): void {
+    if (this.filter.date.after === after && this.filter.date.before === before) return;
+    this.filter.date.after = after;
+    this.filter.date.before = before;
+    this.emit('filter', this.filter);
     this.emit('change');
   }