+++ packages/landing/src/tile.matrix.ts
@@ -1,6 +1,7 @@
 import { GoogleTms, Nztm2000QuadTms, Nztm2000Tms, TileMatrixSet } from '@basemaps/geo';
 import { Projection } from '@basemaps/shared/build/proj/projection.js';
 import { StyleSpecification } from 'maplibre-gl';
+import { DateRangeState } from './components/daterange.js';
 import { Config } from './config.js';
 import { MapLocation, MapOptionType, WindowUrl } from './url.js';
 
@@ -12,8 +13,20 @@ export class TileGrid {
     this.extraZoomLevels = extraZoomLevels;
   }
 
-  getStyle(layerId: string, style?: string | null, config = Config.map.config): StyleSpecification | string {
-    return WindowUrl.toTileUrl(MapOptionType.Style, this.tileMatrix, layerId, style, config);
+  getStyle(
+    layerId: string,
+    style?: string | null,
+    config = Config.map.config,
+    date?: DateRangeState,
+  ): StyleSpecification | string {
+    return WindowUrl.toTileUrl({
+      urlType: MapOptionType.Style,
+      tileMatrix: this.tileMatrix,
+      layerId,
+      style,
+      config,
+      date,
+    });
   }
 }