+++ packages/tiler/src/tiler.ts
@@ -127,8 +127,9 @@ export class Tiler {
     // Often COG tiles do not align to the same size as XYZ Tiles
     // This will scale the COG tile to the same size as a XYZ
     if (source.width !== target.width || source.height !== target.height) {
-      const scale = source.width !== target.width ? target.width / source.width : target.height / source.height;
-      composition.resize = { width: target.width, height: target.height, scale };
+      const scaleX = target.width / source.width;
+      const scaleY = target.height / source.height;
+      composition.resize = { width: target.width, height: target.height, scaleX, scaleY };
     }
 
     // If the output XYZ tile needs a piece of a COG tile, extract the speicific
@@ -139,7 +140,7 @@ export class Tiler {
       tileBounds.height !== target.height ||
       tileBounds.width !== target.width
     ) {
-      composition.crop = tileBounds;
+      composition.crop = tileBounds.toJson();
     }
 
     return composition;